Rewrite workshop support.
System2 may no longer be used since it's max buffer size (4096) was
too much of a limitation, only SteamWorks will work now.

Additionally, the restriction that the workshop maps must already
be on the sever is relaxed and the server with use the host_workshop_map
command to download missing maps.